# perf(errors/gerror): 缓存堆栈信息