--- a/errors/gerror/gerror_error_stack.go
+++ b/errors/gerror/gerror_error_stack.go
@@ -11,10 +11,10 @@ import (
 	"container/list"
 	"fmt"
 	"runtime"
+	"strconv"
 	"strings"
 
 	"github.com/gogf/gf/v2/internal/consts"
-	"github.com/gogf/gf/v2/internal/errors"
 )
 
 // stackInfo manages stack info of certain error.
@@ -35,37 +35,102 @@ func (err *Error) Stack() string {
 	if err == nil {
 		return ""
 	}
-	var (
-		loop             = err
-		index            = 1
-		infos            []*stackInfo
-		isStackModeBrief = errors.IsStackModeBrief()
-	)
-	for loop != nil {
-		info := &stackInfo{
-			Index:   index,
-			Message: fmt.Sprintf("%-v", loop),
+	return getStackInfo(err, 1)
+
+	//var (
+	//	loop             = err
+	//	index            = 1
+	//	infos            []*stackInfo
+	//	isStackModeBrief = errors.IsStackModeBrief()
+	//)
+	//for loop != nil {
+	//	info := &stackInfo{
+	//		Index:   index,
+	//		Message: fmt.Sprintf("%-v", loop),
+	//	}
+	//	index++
+	//	infos = append(infos, info)
+	//	loopLinesOfStackInfo(loop.stack, info, isStackModeBrief)
+	//	if loop.error != nil {
+	//		if e, ok := loop.error.(*Error); ok {
+	//			loop = e
+	//		} else {
+	//			infos = append(infos, &stackInfo{
+	//				Index:   index,
+	//				Message: loop.error.Error(),
+	//			})
+	//			index++
+	//			break
+	//		}
+	//	} else {
+	//		break
+	//	}
+	//}
+	//filterLinesOfStackInfos(infos)
+	//return formatStackInfos(infos)
+}
+
+func getStackInfo(x error, id int) string {
+	switch err := x.(type) {
+	case *Error:
+		prevError := ""
+		if err.error != nil {
+			prevError = "\n" + getStackInfo(err.error, id+1)
 		}
-		index++
-		infos = append(infos, info)
-		loopLinesOfStackInfo(loop.stack, info, isStackModeBrief)
-		if loop.error != nil {
-			if e, ok := loop.error.(*Error); ok {
-				loop = e
-			} else {
-				infos = append(infos, &stackInfo{
-					Index:   index,
-					Message: loop.error.Error(),
-				})
-				index++
-				break
+		var sb strings.Builder
+		count := getIntLength(id) + len(". ") + len(err.text) + len(prevError)
+		pcs := err.stack
+
+		for i, pc := range pcs {
+			count += len("\n\t") + getIntLength(i+1) + len(").")
+
+			buf := cacheStackTrace.get(pc)
+			if buf != nil {
+				count += len(buf)
+				continue
 			}
-		} else {
-			break
+			f, _ := runtime.CallersFrames(pcs[i : i+1]).Next()
+
+			buf = []byte(f.Function + "\n\t\t" + f.File + ":" + strconv.Itoa(f.Line))
+			// TODO: buf变量可能会缓存，全部都是一个变量
+			cacheStackTrace.add(pc, buf)
+			count += len(buf)
 		}
+
+		// 先统计长度，再扩容
+		sb.Grow(count)
+		sb.WriteString(strconv.Itoa(id) + ". " + err.text)
+
+		var b []byte
+		for i, pc := range pcs {
+			b = cacheStackTrace.get(pc)
+			sb.Write(getStackTraceFuncIDHeader(i))
+			sb.Write(b)
+		}
+		if prevError != "" {
+			sb.WriteString(prevError)
+		}
+		return sb.String()
 	}
-	filterLinesOfStackInfos(infos)
-	return formatStackInfos(infos)
+	return x.Error()
+}
+
+// 0-9     = 1
+// 10-99   = 2
+// 100-999 = 3
+// ...
+func getIntLength(n int) int {
+	if n < 10 {
+		return 1
+	}
+	if n < 100 {
+		return 2
+	}
+	if n < 1000 {
+		return 3
+	}
+	// 返回一个超大负数,strings.Builder.Grow() 就会panic
+	return -10000000
 }
 
 // filterLinesOfStackInfos removes repeated lines, which exist in subsequent stacks, from top errors.