--- a/errors/gerror/gerror_error_stack.go
+++ b/errors/gerror/gerror_error_stack.go
@@ -8,55 +8,42 @@ package gerror
 
 import (
 	"bytes"
-	"container/list"
-	"fmt"
+	"math"
 	"runtime"
 	"strconv"
 	"strings"
 	"unsafe"
 
 	"github.com/gogf/gf/v2/internal/consts"
+	"github.com/gogf/gf/v2/internal/errors"
 )
-
-// stackInfo manages stack info of certain error.
-type stackInfo struct {
-	Index   int        // Index is the index of current error in whole error stacks.
-	Message string     // Error information string.
-	Lines   *list.List // Lines contains all error stack lines of current error stack in sequence.
-}
-
-// stackLine manages each line info of stack.
-type stackLine struct {
-	Function string // Function name, which contains its full package path.
-	FileLine string // FileLine is the source file name and its line number of Function.
-}
 
 // Stack returns the error stack information as string.
 func (err *Error) Stack() string {
 	if err == nil {
 		return ""
 	}
-
-	id := 1
-
-	count := 0
-	stackInfos := make([]tempStackInfo, 0, 4)
-
-	tsi := getErrorStackInfo(err, id)
+	var (
+		id               = 1
+		count            = 0
+		stackInfos       = make([]stackInfo, 0, 4)
+		isStackModeBrief = errors.IsStackModeBrief()
+	)
+	tsi := getErrorStackInfo(err, id, isStackModeBrief)
 	stackInfos = append(stackInfos, tsi)
-	count += tsi.count
+	count += int(tsi.count)
 
 	temp := err.error
 	for temp != nil {
 		switch x := temp.(type) {
 		case *Error:
 			id++
-			tsi = getErrorStackInfo(x, id)
+			tsi = getErrorStackInfo(x, id, isStackModeBrief)
 			stackInfos = append(stackInfos, tsi)
-			count += tsi.count + len("\n")
+			count += int(tsi.count) + len("\n")
 			temp = x.error
 		default:
-			// TODO sb.Write(x.Error())
+			// TODO sb.Write(x.Error()) ?
 			break
 		}
 	}
@@ -66,9 +53,9 @@ func (err *Error) Stack() string {
 	defer putBytesBuffer(sb)
 
 	for _, si := range stackInfos {
-		sb.Write(si.funcLine)
+		sb.Write(si.errLine)
 		for k, b := range si.bufptrs {
-			sb.Write(getStackTraceFuncIDHeader(k))
+			sb.Write(getLineIdPrefix(k))
 			sb.Write(*(*[]byte)(b))
 		}
 		sb.WriteByte('\n')
@@ -82,25 +69,26 @@ func (err *Error) stackWithBuffer(buffer *bytes.Buffer, errError string) {
 		return
 	}
 	var (
-		id         = 1
-		count      = 0
-		stackInfos = make([]tempStackInfo, 0, 4)
+		id               = 1
+		count            = 0
+		stackInfos       = make([]stackInfo, 0, 4)
+		isStackModeBrief = errors.IsStackModeBrief()
 	)
-	tsi := getErrorStackInfo(err, id)
-	stackInfos = append(stackInfos, tsi)
-	count += tsi.count
+	errStackInfo := getErrorStackInfo(err, id, isStackModeBrief)
+	stackInfos = append(stackInfos, errStackInfo)
+	count += int(errStackInfo.count) + len("\n")
 
 	temp := err.error
 	for temp != nil {
 		switch x := temp.(type) {
 		case *Error:
 			id++
-			tsi = getErrorStackInfo(x, id)
-			stackInfos = append(stackInfos, tsi)
-			count += tsi.count + len("\n")
+			errStackInfo = getErrorStackInfo(x, id, isStackModeBrief)
+			stackInfos = append(stackInfos, errStackInfo)
+			count += int(errStackInfo.count) + len("\n")
 			temp = x.error
 		default:
-			// TODO buffer.Write(x.Error())
+			// TODO buffer.Write(x.Error()) ?
 			break
 		}
 	}
@@ -114,52 +102,98 @@ func (err *Error) stackWithBuffer(buffer *bytes.Buffer, errError string) {
 	}
 
 	for _, si := range stackInfos {
-		buffer.Write(si.funcLine)
+		buffer.Write(si.errLine)
 		for k, bytesPtr := range si.bufptrs {
-			buffer.Write(getStackTraceFuncIDHeader(k))
+			buffer.Write(getLineIdPrefix(k))
 			buffer.Write(*(*[]byte)(bytesPtr))
 		}
 		buffer.WriteByte('\n')
 	}
 }
 
-type tempStackInfo struct {
-	// buf [][]byte
-	// bufptrs[0] = unsafe.Pointer(&buf[0])
+type stackInfo struct {
+	// bufs [][]byte
+	// bufptrs[0] = unsafe.Pointer(&bufs[0])
+	// bufptrs[1] = unsafe.Pointer(&bufs[1])
+	// ...
 	bufptrs []unsafe.Pointer
-	// funcName
-	// 		xx/yy/zz:10
-	funcLine []byte
-	// count+=len [for len range len(buf[i])]
-	count int
+	// 1. error1
+	// 2. error2
+	// ...
+	errLine []byte
+	// for _,buf:=range len(bufs) {
+	//		count += len(buf)
+	//}
+	// Calculate the space that needs to be expanded to avoid further expansion during writing
+	count int32
 }
 
-func getErrorStackInfo(err *Error, id int) (tsi tempStackInfo) {
+func getErrorStackInfo(err *Error, id int, isStackModeBrief bool) (errStackInfo stackInfo) {
 	var (
 		count = getIntLength(id) + len(". ") + len(err.text)
 		pcs   = err.stack
 	)
-	tsi.bufptrs = make([]unsafe.Pointer, len(pcs))
+	errStackInfo.bufptrs = make([]unsafe.Pointer, 0, len(pcs))
 
 	for i, pc := range pcs {
-		count += len("\n\t") + getIntLength(i+1) + len(").")
+		// count += len("\n\t%d). ")
+		count += len("\n\t") + getIntLength(i+1) + len("). ")
 
-		bufptr := cacheStackTrace.getBytesPtr(pc)
-		if bufptr != nil {
-			count += len(*bufptr)
-			tsi.bufptrs[i] = unsafe.Pointer(bufptr)
-			continue
+		bufptr, flag := cacheStackTrace.getBytesPtr(pc)
+		if bufptr == nil {
+			bufptr, flag = addStackInfoToCache(pc)
 		}
+		count += len(*bufptr)
 
-		f, _ := runtime.CallersFrames(pcs[i : i+1]).Next()
-		buf := []byte(f.Function + "\n\t\t" + f.File + ":" + strconv.Itoa(f.Line))
-		bufptr = cacheStackTrace.addAndGetBytesPtr(pc, buf)
-		count += len(buf)
-		tsi.bufptrs[i] = unsafe.Pointer(bufptr)
+		// 1. If it is in brief mode and includes the path of the [goframe] package, skip it directly
+		//    isStackBrief && flag == stackObjectFlagStackFilterKeyForGoFrame
+		// 2. Path containing [gerror] package, skip
+		//    flag == stackObjectFlagStackFilterKeyLocal
+		// TODO: Ignore the automatically generated package path and goroot path?
+		// ignore autogenerated<>  ?
+		// ignore go root paths    ?
+
+		if isStackModeBrief {
+			if flag == stackObjectFlagStackFilterKeyForGoFrame {
+				continue
+			}
+			errStackInfo.bufptrs = append(errStackInfo.bufptrs, unsafe.Pointer(bufptr))
+		} else if flag == stackObjectFlagStackFilterKeyLocal {
+			// continue
+		} else {
+			errStackInfo.bufptrs = append(errStackInfo.bufptrs, unsafe.Pointer(bufptr))
+		}
+		// ignore autogenerated<>  ?
+		// ignore go root paths    ?
 	}
-	tsi.funcLine = []byte(strconv.Itoa(id) + ". " + err.text)
-	tsi.count = count
-	return tsi
+	errStackInfo.errLine = []byte(getIdStr(id) + ". " + err.text)
+	errStackInfo.count = int32(count)
+	return errStackInfo
+}
+
+func addStackInfoToCache(pc uintptr) (*[]byte, stackObjectFlag) {
+	pcs := [1]uintptr{pc}
+	f, _ := runtime.CallersFrames(pcs[:]).Next()
+	buf := []byte(f.Function + "\n\t\t" + f.File + ":" + strconv.Itoa(f.Line))
+
+	//fn = runtime.FuncForPC(pc - 1)
+	//file, line = fn.FileLine(pc - 1)
+	//buf := []byte(fn.Name() + "\n\t\t" + file + ":" + strconv.Itoa(line))
+
+	const gf_path = consts.StackFilterKeyForGoFrame
+	flag := stackObjectFlag(0)
+	if strings.Contains(f.File, gf_path) {
+		// package path stack filtering.
+		if strings.Contains(f.File, stackFilterKeyLocal) {
+			//continue
+			flag = stackObjectFlagStackFilterKeyLocal
+		} else {
+			//continue
+			flag = stackObjectFlagStackFilterKeyForGoFrame
+		}
+	}
+	bufptr := cacheStackTrace.addAndGetBytesPtr(pc, buf, flag)
+	return bufptr, flag
 }
 
 // 0-9     = 1
@@ -176,110 +210,11 @@ func getIntLength(n int) int {
 	if n < 1000 {
 		return 3
 	}
-	// 返回一个超大负数,strings.Builder.Grow() 就会panic
-	return -10000000
-}
-
-// filterLinesOfStackInfos removes repeated lines, which exist in subsequent stacks, from top errors.
-func filterLinesOfStackInfos(infos []*stackInfo) {
-	var (
-		ok      bool
-		set     = make(map[string]struct{})
-		info    *stackInfo
-		line    *stackLine
-		removes []*list.Element
-	)
-	for i := len(infos) - 1; i >= 0; i-- {
-		info = infos[i]
-		if info.Lines == nil {
-			continue
-		}
-		for n, e := 0, info.Lines.Front(); n < info.Lines.Len(); n, e = n+1, e.Next() {
-			line = e.Value.(*stackLine)
-			if _, ok = set[line.FileLine]; ok {
-				removes = append(removes, e)
-			} else {
-				set[line.FileLine] = struct{}{}
-			}
-		}
-		if len(removes) > 0 {
-			for _, e := range removes {
-				info.Lines.Remove(e)
-			}
-		}
-		removes = removes[:0]
+	if n < 10000 {
+		return 4
 	}
-}
-
-// formatStackInfos formats and returns error stack information as string.
-func formatStackInfos(infos []*stackInfo) string {
-	buffer := bytes.NewBuffer(nil)
-	for i, info := range infos {
-		buffer.WriteString(fmt.Sprintf("%d. %s\n", i+1, info.Message))
-		if info.Lines != nil && info.Lines.Len() > 0 {
-			formatStackLines(buffer, info.Lines)
-		}
+	if n < 100000 {
+		return 5
 	}
-	return buffer.String()
-}
-
-// formatStackLines formats and returns error stack lines as string.
-func formatStackLines(buffer *bytes.Buffer, lines *list.List) string {
-	var (
-		line   *stackLine
-		space  = "  "
-		length = lines.Len()
-	)
-	for i, e := 0, lines.Front(); i < length; i, e = i+1, e.Next() {
-		line = e.Value.(*stackLine)
-		// Graceful indent.
-		if i >= 9 {
-			space = " "
-		}
-		buffer.WriteString(fmt.Sprintf(
-			"   %d).%s%s\n        %s\n",
-			i+1, space, line.Function, line.FileLine,
-		))
-	}
-	return buffer.String()
-}
-
-// loopLinesOfStackInfo iterates the stack info lines and produces the stack line info.
-func loopLinesOfStackInfo(st stack, info *stackInfo, isStackModeBrief bool) {
-	if st == nil {
-		return
-	}
-	for _, p := range st {
-		if fn := runtime.FuncForPC(p - 1); fn != nil {
-			file, line := fn.FileLine(p - 1)
-			if isStackModeBrief {
-				// filter whole GoFrame packages stack paths.
-				if strings.Contains(file, consts.StackFilterKeyForGoFrame) {
-					continue
-				}
-			} else {
-				// package path stack filtering.
-				if strings.Contains(file, stackFilterKeyLocal) {
-					continue
-				}
-			}
-			// Avoid stack string like "`autogenerated`"
-			if strings.Contains(file, "<") {
-				continue
-			}
-			// Ignore GO ROOT paths.
-			if goRootForFilter != "" &&
-				len(file) >= len(goRootForFilter) &&
-				file[0:len(goRootForFilter)] == goRootForFilter {
-				continue
-			}
-			if info.Lines == nil {
-				info.Lines = list.New()
-			}
-			info.Lines.PushBack(&stackLine{
-				Function: fn.Name(),
-				FileLine: fmt.Sprintf(`%s:%d`, file, line),
-			})
-		}
-	}
+	return int(math.Log10(float64(n)) + 1)
 }