为gerror缓存堆栈信息，缓存之后，性能提升几倍，内存申请也减少了许多，不过目前仍有一些问题与主分支的代码输出的结果不同，
gerror类型使用Wrap再次包装之后，再次输出的时候，两个错误的堆栈信息合并到一起了，这样看起来能够大幅度减少一些噪音，但似乎Wrap之后的error堆栈信息丢失了，只能看到一个“gerror/Wrap包装数据3”错误信息，这样是否不妥？
我使用"github.com/pkg/errors"这个包，使用相同的逻辑，使用errors.New，然后在Test3函数重新使用errors.Wrap包装一下，再次返回，输出的结果也是两次堆栈信息都是独立输出的，并没有合并到一起去

此外本次提交的代码没有忽略标准库的路径还有编译器自动生成的代码路径
```go
func benchGfErrorsTest1() error {
	return benchGfErrorsTest2()
}

func benchGfErrorsTest2() error {
	return benchGfErrorsTest3()
}

func benchGfErrorsTest3() error {
	return gerror.Wrap(benchGfErrorsTest4(), "gerror/Wrap包装数据3")
}

func benchGfErrorsTest4() error {
	return benchGfErrorsTest5()
}

func benchGfErrorsTest5() error {
	return gerror.New("gerror/New测试数据")
}

func main() {
	err := benchGfErrorsTest1()
	fmt.Printf("%+v\n", err)
}

运行上面代码

```

## gf的输出结果为
```go
gerror/Wrap包装数据3: gerror/New测试数据
1. gerror/Wrap包装数据3
2. gerror/New测试数据
   1).  main.benchGfErrorsTest5
        go-test/test.go:63
   2).  main.benchGfErrorsTest4
        go-test/test.go:59
   3).  main.benchGfErrorsTest3
        go-test/test.go:55
   4).  main.benchGfErrorsTest2
        go-test/test.go:51
   5).  main.benchGfErrorsTest1
        go-test/test.go:47
   6).  main.main
        go-test/gerror.go:6
```

## 此次pr的输出结果
```go
gerror/Wrap包装数据3: gerror/New测试数据
1. gerror/Wrap包装数据3
        1). main.benchGfErrorsTest3
                go-test/test.go:55
        2). main.benchGfErrorsTest2
                go-test/test.go:51
        3). main.benchGfErrorsTest1
                go-test/test.go:47
        4). main.main
                go-test/gerror.go:6
        5). runtime.main
                I:/go1.22/src/runtime/proc.go:271
        6). runtime.goexit
                I:/go1.22/src/runtime/asm_amd64.s:1695
2. gerror/New测试数据
        1). main.benchGfErrorsTest5
                go-test/test.go:63
        2). main.benchGfErrorsTest4
                go-test/test.go:59
        3). main.benchGfErrorsTest3
                go-test/test.go:55
        4). main.benchGfErrorsTest2
                go-test/test.go:51
        5). main.benchGfErrorsTest1
                go-test/test.go:47
        6). main.main
                go-test/gerror.go:6
        7). runtime.main
                I:/go1.22/src/runtime/proc.go:271
        8). runtime.goexit
                I:/go1.22/src/runtime/asm_amd64.s:1695
```
## 性能比对
分为直接用new返回，中间不做任何处理，还有一个是中间用Wrap包装了下new返回的error
都使用%+v解开堆栈信息来测试
```go
// 测试命令 go test -v -run=none -bench=^Benchmark_gfErrors -benchtime=5s -benchmem -cpuprofile='cpu.out' -memprofile='mem.out' -count=3 .
func Benchmark_gfErrors(b *testing.B) {
	b.ReportAllocs()
	var err error
	s := " "
	for range b.N {
		err = benchGfErrorsTest1()
		s = fmt.Sprintf("%+v", err)
	}
	_ = s
}

func Benchmark_gfErrors_Parallel(b *testing.B) {
	b.ReportAllocs()
	var err error
	s := " "
	b.RunParallel(func(p *testing.PB) {
		for p.Next() {
			err = benchGfErrorsTest1()
			s = fmt.Sprintf("%+v", err)
		}
	})
	_ = s
}
```

### 当前主分支的new
```go
Benchmark_gfErrors
Benchmark_gfErrors-8            	  320312	     15957 ns/op	    8040 B/op	      70 allocs/op
Benchmark_gfErrors-8            	  359022	     15981 ns/op	    8040 B/op	      70 allocs/op
Benchmark_gfErrors-8            	  361665	     15617 ns/op	    8040 B/op	      70 allocs/op
Benchmark_gfErrors_Parallel
Benchmark_gfErrors_Parallel-8   	 1000000	      5350 ns/op	    8185 B/op	      70 allocs/op
Benchmark_gfErrors_Parallel-8   	 1000000	      5108 ns/op	    8184 B/op	      70 allocs/op
Benchmark_gfErrors_Parallel-8   	 1000000	      5140 ns/op	    8183 B/op	      70 allocs/op
```
### 当前主分支的wrap
```go
Benchmark_gfErrors
Benchmark_gfErrors-8            	  256899	     24295 ns/op	    8954 B/op	     101 allocs/op
Benchmark_gfErrors-8            	  249151	     24198 ns/op	    8954 B/op	     101 allocs/op
Benchmark_gfErrors-8            	  264451	     23054 ns/op	    8953 B/op	     101 allocs/op
Benchmark_gfErrors_Parallel
Benchmark_gfErrors_Parallel-8   	  783068	      7812 ns/op	   10508 B/op	     102 allocs/op
Benchmark_gfErrors_Parallel-8   	  694832	      7815 ns/op	   10506 B/op	     102 allocs/op
Benchmark_gfErrors_Parallel-8   	  872317	      7450 ns/op	   10506 B/op	     102 allocs/op
```
### 当前pr的new
```go
Benchmark_gfErrors
Benchmark_gfErrors-8            	 1912519	      3117 ns/op	    1794 B/op	       6 allocs/op
Benchmark_gfErrors-8            	 1857090	      3092 ns/op	    1794 B/op	       6 allocs/op
Benchmark_gfErrors-8            	 1974368	      3112 ns/op	    1794 B/op	       6 allocs/op
Benchmark_gfErrors_Parallel
Benchmark_gfErrors_Parallel-8   	 6405152	      1009 ns/op	    1761 B/op	       6 allocs/op
Benchmark_gfErrors_Parallel-8   	 6406886	       966.3 ns/op	    1761 B/op	       6 allocs/op
Benchmark_gfErrors_Parallel-8   	 5790344	       951.7 ns/op	    1761 B/op	       6 allocs/op
```
### 当前pr的wrap
```go
Benchmark_gfErrors
Benchmark_gfErrors-8            	 1000000	      5614 ns/op	    3366 B/op	      13 allocs/op
Benchmark_gfErrors-8            	 1000000	      5713 ns/op	    3366 B/op	      13 allocs/op
Benchmark_gfErrors-8            	 1000000	      5631 ns/op	    3366 B/op	      13 allocs/op
Benchmark_gfErrors_Parallel
Benchmark_gfErrors_Parallel-8   	 3922666	      1579 ns/op	    3172 B/op	      13 allocs/op
Benchmark_gfErrors_Parallel-8   	 3632580	      1635 ns/op	    3172 B/op	      13 allocs/op
Benchmark_gfErrors_Parallel-8   	 3618468	      1620 ns/op	    3172 B/op	      13 allocs/op
```