@gqcn 2024-12-23
@wln32 
> gerror类型使用Wrap再次包装之后，再次输出的时候，两个错误的堆栈信息合并到一起了，这样看起来能够大幅度减少一些噪音，但似乎Wrap之后的error堆栈信息丢失了，只能看到一个“gerror/Wrap包装数据3”错误信息，这样是否不妥？

这块是`goframe`的错误处理组件特意做的降噪处理，因为`Wrap`后的`error`有重复的代码行，特别是针对多次`Warp`后层级比较深的`error`，效果比较明显，也不影响问题定位。


> 关于`gerror`的性能改进。

由于`error`只在错误产生时才会执行调用，虽然生成堆栈的性能会较差，但是也不是什么问题。缓存这版实现我没有看明白，并且`gerror`是核心组件，建议类似改动可以先提交`draft`描述思路设置设计方案，再去落地代码实现。