@wln32 2024-12-23
> @wln32
> 
> > gerror类型使用Wrap再次包装之后，再次输出的时候，两个错误的堆栈信息合并到一起了，这样看起来能够大幅度减少一些噪音，但似乎Wrap之后的error堆栈信息丢失了，只能看到一个“gerror/Wrap包装数据3”错误信息，这样是否不妥？
> 
> 这块是`goframe`的错误处理组件特意做的降噪处理，因为`Wrap`后的`error`有重复的代码行，特别是针对多次`Warp`后层级比较深的`error`，效果比较明显，也不影响问题定位。
> 
> > 关于`gerror`的性能改进。
> 
> 由于`error`只在错误产生时才会执行调用，虽然生成堆栈的性能会较差，但是也不是什么问题。缓存这版实现我没有看明白，并且`gerror`是核心组件，建议类似改动可以先提交`draft`描述思路设置设计方案，再去落地代码实现。


@gqcn 
> > 这块是`goframe`的错误处理组件特意做的降噪处理，因为`Wrap`后的`error`有重复的代码行，特别是针对多次`Warp`后层级比较深的`error`，效果比较明显，也不影响问题定位。

这样还是有影响的吧，只能看到是两个错误，然后是堆栈信息，但是并不知道哪个错误的堆栈在哪里，还得一个个看

> > 由于`error`只在错误产生时才会执行调用，虽然生成堆栈的性能会较差，但是也不是什么问题

不能认同

> > 缓存这版实现我没有看明白

大概思路就是使用先前保存的pc指针，也就是每一层调用栈函数的pc的指针，
1. 然后调用runtime.CallersFrames(pcs[:]).Next()来获取函数的名字，行号，文件名等信息，
这一步会做缓存，先根据pc指针在缓存中查找，如果找到了，就不用在调用runtime库重复获取，这个api调用比较耗时。
如果没找到，再调用api将其添加到缓存中
2. 使用bytes.Buffer或者strings.Builder这样的带缓冲的结构来保存本次需要打印的信息，如果在调用Write时发生了扩容，这也会比较耗时，所以这部分的思路就是用一个count变量统计本次需要打印的堆栈信息的长度，直接一次扩容，确保不会再写入时再次发生扩容，直接写入即可
3. 然后就是把每次使用的bytes.Buffer放进sync.Pool来做缓存，也可以减少gc或者内存申请的次数

> > 建议类似改动可以先提交`draft`描述思路设置设计方案，再去落地代码实现。

其实这次pr的想法是我看到了一篇文章，里面介绍了可以把堆栈的一些信息给保存起来，再次使用的时候，不需要再次调用runtime库获取，我在本地实现了一下，代码其实也不多，就四五十行左右（本地那版实现比现在的性能还要高出两倍），我就想尝试在gerror看能不能实现下，目前来说，可以实现的和先前的效果一模一样，没什么难度
这种思路其实还可以用在其他地方，像glog，需要打印文件名函数名行号的，也都可以做到获取一次，后续重复使用，几乎没有什么成本