@gqcn 2024-12-25
@wln32 堆栈降噪那个是很有必要的，因为`Wrap`后的新`error`必然会存在重复堆栈，顶部的错误会包含所有的堆栈。至于性能那个，我觉得思路挺不错的，这个思路也很简单。但是这个实现的代码，老实说，有点看不明白啊，我花很多时间去`review`都看不明白，我有点头疼。能否让AI来优化一下代码，使得实现代码更优雅易维护一些。